+++ ft_check.c
@@ -6,7 +6,7 @@
 /*   By: lprior <lprior@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/07 18:26:23 by lprior            #+#    #+#             */
-/*   Updated: 2018/02/09 15:04:44 by lprior           ###   ########.fr       */
+/*   Updated: 2018/02/09 17:50:38 by lprior           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,7 +45,7 @@
 // #define RETURN_J	T ? (intmax_t)ARG : (uintmax_t)ARG
 // #define RETURN_Z	(size_t)va_arg(ap, size_t)
 
-int ft_check_signed_int(t_flags *tools, va_list ap)
+long long int ft_check_signed_int(t_flags *tools, va_list ap)
 {
     if (tools->brand == 'd' || tools->brand == 'i')
 		return ((long long int)va_arg(ap, void*));
@@ -58,12 +58,12 @@ unsigned long long int ft_check_unsigned_int(t_flags *tools, va_list ap)
 		return ((unsigned long long int)va_arg(ap, void*));
     else if (tools->brand == 'u' || tools->brand == 'o' || tools->brand == 'X' 
             || tools->brand == 'x' || tools->brand == 'D')
-		return ((unsigned int)va_arg(ap, void*)); 
+		return ((unsigned long long int)va_arg(ap, void*)); 
     return ((unsigned long long int)NULL);
 }
 
 
-long long int ft_sort_unsigned_args(t_flags *tools, va_list ap)
+unsigned long long int ft_sort_unsigned_args(t_flags *tools, va_list ap)
 {
     if (tools->arg == 1)
         return ((size_t)va_arg(ap, size_t));
